+++ org.lxtk/src/org/lxtk/client/RenameFeature.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2020, 2021 1C-Soft LLC.
+ * Copyright (c) 2020, 2022 1C-Soft LLC.
  *
  * This program and the accompanying materials are made available under
  * the terms of the Eclipse Public License 2.0 which is available at
@@ -20,6 +20,7 @@ import java.util.UUID;
 import java.util.concurrent.CompletableFuture;
 
 import org.eclipse.lsp4j.DocumentFilter;
+import org.eclipse.lsp4j.PrepareRenameDefaultBehavior;
 import org.eclipse.lsp4j.PrepareRenameParams;
 import org.eclipse.lsp4j.PrepareRenameResult;
 import org.eclipse.lsp4j.Range;
@@ -31,6 +32,7 @@ import org.eclipse.lsp4j.ServerCapabilities;
 import org.eclipse.lsp4j.TextDocumentClientCapabilities;
 import org.eclipse.lsp4j.WorkspaceEdit;
 import org.eclipse.lsp4j.jsonrpc.messages.Either;
+import org.eclipse.lsp4j.jsonrpc.messages.Either3;
 import org.lxtk.LanguageService;
 import org.lxtk.ProgressService;
 import org.lxtk.RenameProvider;
@@ -127,8 +129,9 @@ public class RenameFeature
             }
 
             @Override
-            public CompletableFuture<Either<Range, PrepareRenameResult>> prepareRename(
-                PrepareRenameParams params)
+            public CompletableFuture<
+                Either3<Range, PrepareRenameResult, PrepareRenameDefaultBehavior>> prepareRename(
+                    PrepareRenameParams params)
             {
                 if (!Boolean.TRUE.equals(options.getPrepareProvider()))
                     throw new UnsupportedOperationException();